requirements.txt
redis == 2.10.3
requests == 2.7.0
selenium == 2.48.0
SQLAlchemy == 1.0.8
twython == 3.3.0
Pillow == 3.0.0
